requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
executing==2.0.1
Flask==3.0.3
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
icecream==2.1.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.30.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pyngrok==7.1.6
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
